requirements.txt
openai
websockets
python-dotenv
